bootloader::boot_info

Enum MemoryRegionKind

#[non_exhaustive]

#[repr(C)]
pub enum MemoryRegionKind {
    Usable,
    Bootloader,
    UnknownUefi(u32),
    UnknownBios(u32),
}

Represents the different types of memory.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Usable

Unused conventional memory, can be used by the kernel.

Bootloader

Memory mappings created by the bootloader, including the kernel and boot info mappings.

This memory should not be used by the kernel.

UnknownUefi(u32)

An unknown memory region reported by the UEFI firmware.

This should only be used if the UEFI memory type is known as usable.

UnknownBios(u32)

An unknown memory region reported by the BIOS firmware.

This should only be used if the BIOS memory type is known as usable.