bootloader::boot_info

Enum Optional

#[repr(C)]
pub enum Optional<T> {
    Some(T),
    None,
}

FFI-safe variant of [Option].

Implements the [From] and [Into] traits for easy conversion to and from [Option].

Variants

Some(T)

Some value T

None

No value

Implementations

impl<T> Optional<T>

pub fn into_option(self) -> Option<T>

Converts the Optional to an [Option].

pub const fn as_ref(&self) -> Option<&T>

Converts from &Optional<T> to Option<&T>.

For convenience, this method directly performs the conversion to the standard [Option] type.

pub fn as_mut(&mut self) -> Option<&mut T>

Converts from &mut Optional<T> to Option<&mut T>.

For convenience, this method directly performs the conversion to the standard [Option] type.