bootloader::boot_info

Struct TlsTemplate

#[repr(C)]
pub struct TlsTemplate {
    pub start_addr: u64,
    pub file_size: u64,
    pub mem_size: u64,
}

Information about the thread local storage (TLS) template.

This template can be used to set up thread local storage for threads. For each thread, a new memory location of size mem_size must be initialized. Then the first file_size bytes of this template needs to be copied to the location. The additional mem_size - file_size bytes must be initialized with zero.

Fields
start_addr: u64

The virtual start address of the thread local storage template.

file_size: u64

The number of data bytes in the template.

Corresponds to the length of the .tdata section.

mem_size: u64

The total number of bytes that the TLS segment should have in memory.

Corresponds to the combined length of the .tdata and .tbss sections.