Crate bootloader

The bootloader is comprised of two main stages, along with a short initialization stage within the reset handler.

Reset Handler

The reset handler is provided by the cortex_m_rt crate and resides in flash at 0x5800. This is the very first thing that executes (barring MITRE bootstrapper code).

Implementation Details

The stack pointer is reinitialized.

The .data section is copied from flash to SRAM.

The .bss section is zeroed in SRAM.

stage1 is executed.

Stage 1

The first stage is the initial entry point to the bootloader and is responsible for setting up the second stage See stage1 for more.

Stage 2

The second stage contains the core logic for interacting with the host-tools. See stage2 for more.

Functions

__cortex_m_rt_main

Initial entry point. This immediately delegates to stage1.

decrypt_stage2

Decrypts the second stage.

reset_sp⚠

Resets the stack pointer to the end of available SRAM.

run_cmd

Blocks until a command is received from host-tools, then executes the respective handler. Invalid commands are rejected.

stage1

The initial execution stage of the bootloader.

stage2

The chief execution stage of the bootloader.