Module bootloader::boot_info

Contains the boot information struct sent by the bootloader to the kernel on startup.

Structs

BootInfo
This structure represents the information that the bootloader passes to the kernel.

FrameBuffer
A pixel-based framebuffer that controls the screen output.

FrameBufferInfo
Describes the layout and pixel format of a framebuffer.

MemoryRegion
Represent a physical memory region.

MemoryRegions
FFI-safe slice of MemoryRegion structs, semantically equivalent to &'static mut [MemoryRegion].

TlsTemplate
Information about the thread local storage (TLS) template.

Enums

MemoryRegionKind
Represents the different types of memory.

Optional
FFI-safe variant of Option.

PixelFormat
Color format of pixels in the framebuffer.